gateway/requirements.txt
flask
requests
flask_jwt_extended
pandas